+++ components/nutshell/CHANGELOG.md
@@ -4,8 +4,19 @@
 
 ## [0.1.1] - 2026-02-24
 
+### Added
+
+- update-company action
+- update-contact action
+- update-lead action
+- search-company action
+- search-contact action
+- search-lead action
+- get-company action
+- get-contact action
+- get-lead action
+
 ### Changed
 
-- Added update-company, update-contact, update-lead, search-company, search-contact, search-lead, get-company, get-contact, get-lead functionality.
 - Improved lead, contact, and company output schemas
 - App-level improvements in nutshell.app.mjs